📝 document scope, spawn
Spiff up the scope documentation. The `ScopeInternal` class was moved
to its own file so that it can be imported internally, but is not
exported from the main module.
